# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new group",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/members": {
      "post": {
        "operationId": "GroupsController_addMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member added to group successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a member to a group (by groupId UUID)",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/validate-statement-of-work": {
      "get": {
        "operationId": "GroupsController_validateSow",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check SoW status for a group",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/committee": {
      "get": {
        "operationId": "getCommitteeByGroupId",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committee found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitteeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Authenticated but forbidden by policy"
          },
          "404": {
            "description": "Group not found or no committee assigned"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the committee assigned to a group (any authenticated user)",
        "tags": [
          "Groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Senior App API",
    "description": "OpenAPI spec generated from backend routes",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "groupName": {
            "type": "string",
            "example": "Student project group",
            "description": "Name of the group"
          },
          "leaderUserId": {
            "type": "string",
            "example": "69e22c0a65cf84775acc127e",
            "description": "MongoDB ObjectId of the group leader user"
          }
        },
        "required": [
          "groupName",
          "leaderUserId"
        ]
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "memberUserId": {
            "type": "string",
            "example": "69e22c0a65cf84775acc127e",
            "description": "MongoDB ObjectId of the user to add to the group"
          }
        },
        "required": [
          "memberUserId"
        ]
      },
      "JuryMemberResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name"
        ]
      },
      "CommitteeAdvisorResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name"
        ]
      },
      "CommitteeGroupResponse": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "description": "UUID of the assigned group"
          },
          "assignedAt": {
            "type": "string",
            "description": "Timestamp when this group was assigned",
            "format": "date-time"
          },
          "assignedByUserId": {
            "type": "string",
            "description": "ID of the user who made this assignment"
          }
        },
        "required": [
          "groupId",
          "assignedAt",
          "assignedByUserId"
        ]
      },
      "CommitteeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID of the committee"
          },
          "name": {
            "type": "string",
            "description": "Name of the committee"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "object",
            "description": "Last update timestamp",
            "nullable": true
          },
          "jury": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JuryMemberResponse"
            }
          },
          "advisors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitteeAdvisorResponse"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitteeGroupResponse"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "jury",
          "advisors",
          "groups"
        ]
      },
...
    }
  }
}